refactor: MarkdownRenderer Swift 연결 경계 축소
