+++ Application/Presentation/PresentationShared/Sources/Common/TodoMarkdownContentView.swift
@@ -9,10 +9,6 @@ import SwiftUI
 import Domain
 
 struct TodoMarkdownContentView: View {
-    @Environment(\.colorScheme) private var colorScheme
-    @Environment(\.locale) private var locale
-    @Environment(\.openURL) private var openURL
-    @ScaledMetric(relativeTo: .body) private var fontSize = 17
     @State private var tabBarHeight = CGFloat.zero
 
     let content: String
@@ -23,12 +19,8 @@ struct TodoMarkdownContentView: View {
         MarkdownRendererView(
             markdown: content,
             references: rendererReferences,
-            colorScheme: colorScheme,
-            languageCode: locale.language.languageCode?.identifier ?? "und",
-            fontSize: fontSize,
             obscuredBottomInset: tabBarHeight,
-            onOpenTodoID: onOpenTodoID,
-            onOpenURL: { openURL($0) }
+            onOpenTodoID: onOpenTodoID
         )
         .frame(maxWidth: .infinity, maxHeight: .infinity)
         .ignoresSafeArea(.container, edges: ignoredSafeAreaEdges)
